refactor UpdateNodeFragment & CreateNodeFragment
